refactor(auth): extract common OAuth handler logic into a reusable utility
